Move converse-bookmarks plugin into folder
